tbnow-back/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
sentence-transformers
faiss-cpu
pypdf
opencv-python
numpy
matplotlib
scikit-learn
dotenv
google-genai
pillow
opencv-python